// src/app/core/shared/models/cart-item-vendor.model.ts
export interface cartItemVendor{
    cartitem_id: number | any;
    product_id: number | any;
    vendor_id: number | any;
    order_id: number | any;
    seller_id: number | any;
    product_name: string | any;
    price: number | any;
    stock: number | any;
    quantity: number | any;
    total: string | any;
    path: string | any;
}